POST method of adding product to the favourite products list finished.
GET method that shows to the user his own favourite products (for now, only if he clicks the products and he's visualizing the card of the individual product) done. Still have to do that the user can watch if it's a favourite product or not in the view where he visualizes all the products.

!!! DELETE method not working because auth token and params don't work, still have to review it.

fix: button now appear as white by default if not added as favourite.